requirements.txt
piexif==1.1.3
Pillow==11.2.1
